Move FlatArray impl into flat_array.rs
